UnifiedCRMIntegration/requirements.txt
Flask==2.3.3
Flask-SQLAlchemy==3.0.5
SQLAlchemy==2.0.21
Flask-Migrate==4.0.5
Flask-RESTX==1.3.0
jsonschema==4.19.1
python-dotenv==1.0.0
psycopg2-binary==2.9.7
Werkzeug==2.3.7
redis==5.0.1
requests==2.31.0
alembic==1.12.0
gunicorn==21.2.0
PyJWT==2.8.0